backend/pyproject.toml
[project]
name = "inrtu-web-django"
version = "1.0.0"
description = "web app based on django and vue.js implemented during INRTU studying"
authors = [
    {name = "Timur Semennikov",email = "wqqzxts@gmail.com"}
]
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    "django (==5.1.1)",
    "djangorestframework (==3.15.2)",
    "pillow (==12.0.0)",
    "uvicorn (==0.38.0)",
    "pytest-django (==4.9.0)",
    "model-bakery (==1.20.0)",
    "whitenoise (==6.11.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "psutil (>=7.1.3,<8.0.0)"
]

[dependency-groups]
dev = [
    "faker (==13.1.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
